requirements.txt
discord_webhook
python-dotenv
influxdb_client
pandas
matplotlib
cloudinary